feat: 동시성 제어 및 캐싱 최적화 브랜치 통합

+++ src/main/java/org/hyejoon/cuvcourse/domain/lecture/lectureView/dto/LectureViewResponse.java
@@ -1,5 +1,6 @@
 package org.hyejoon.cuvcourse.domain.lecture.lectureView.dto;
 
 public record LectureViewResponse(Long id, String lectureTitle, String professorName, int credits,
-                                  int totalStudents, int capacity) {
+                                  long totalStudents, int capacity) {
+
 }
